Modified the date creation of both todo lists and todo list items to be of type long instead of integer. The database later on will store it as an integer.

+++ app/src/main/java/todo/javier/mera/todolist/fragments/FragmentHome.java
@@ -139,7 +139,7 @@ public class FragmentHome extends Fragment
                     mPresenter.updateEditTextHintColor(mParent, android.R.color.darker_gray);
 
                     TodoListDataSource dataSource = new TodoListDataSource(mParent);
-                    int creationDate = (int)new Date().getTime();
+                    long creationDate = new Date().getTime();
 
                     dataSource.openWriteable();
                     long newId = dataSource.createTodoList(
